Check in current state of new ingest framework for safekeeping

+++ Core/src/org/sleuthkit/autopsy/ingest/IngestScheduler.java
@@ -459,8 +459,7 @@ class IngestScheduler {
                 this.dataSourceTask = dataSourceTask;
             }
             
-            
-            void execute(int threadId) {
+            void execute(long threadId) {
                 dataSourceTask.getIngestPipelines().ingestFile(threadId, file);
             }
